requirements.txt
openai>=1.13.3
python-dotenv>=1.0.1
SpeechRecognition>=3.10.0
pyaudio>=0.2.13 ; platform_system == "Windows"
pyaudio>=0.2.13 ; platform_system != "Windows"
pyttsx3>=2.90
elevenlabs>=1.3.1
flask>=3.0.2
flask-socketio>=5.3.6
gpiozero>=1.6.2 ; platform_system == "Linux"
RPi.GPIO>=0.7.1 ; platform_system == "Linux"
requests>=2.31.0
